fix: cors preflight 404 error
For an unknown reason, with bun version, http module seem to keep passing method as uppercase1.0.10, but not options which become lowercase. find-my-way stores routes in a variable this.trees which contain uppercase http methods properties, and so all OPTIONS routes are broken.

Issue: https://github.com/oven-sh/bun/issues/7250